+++ twcore/twcore/core/PlayerEntered.java
@@ -2,58 +2,58 @@ package twcore.core;
 
 public class PlayerEntered extends SubspaceEvent {
 
-    int             m_team;
-    int             m_wins;
-    int             m_losses;
-    int             m_hasKOTH;
-    int             m_shipType;
+    short           m_team;
+    short           m_wins;
+    short           m_losses;
+    byte            m_hasKOTH;
+    byte            m_shipType;
     String          m_squadName;
     String          m_playerName;
     int             m_flagPoints;
     int             m_killPoints;
-    int             m_playerID;
-    int             m_acceptsAudio;
-    int             m_flagsCarried;
-    int             m_identTurretee;
+    short           m_playerID;
+    byte            m_acceptsAudio;
+    short           m_flagsCarried;
+    short           m_identTurretee;
 
     public PlayerEntered( ByteArray array ){
         
-        m_shipType = ((int)array.readByte( 1 ) + 1) % 9;
-        m_acceptsAudio = (int)array.readByte( 2 );
+        m_shipType = (byte)((array.readByte( 1 ) + 1) % 9);
+        m_acceptsAudio = array.readByte( 2 );
         m_playerName = array.readString( 3, 20 );
         m_squadName = array.readString( 23,20 );
         m_flagPoints = array.readLittleEndianInt( 43 );
         m_killPoints = array.readLittleEndianInt( 47 );
-        m_playerID = (int)array.readLittleEndianShort( 51 );
-        m_team = (int)array.readLittleEndianShort( 53 );
-        m_wins = (int)array.readLittleEndianShort( 55 );
-        m_losses = (int)array.readLittleEndianShort( 57 );
-        m_identTurretee = (int)array.readLittleEndianShort( 59 );
-        m_flagsCarried = (int)array.readLittleEndianShort( 61 );
-        m_hasKOTH = (int)array.readByte( 63 );
+        m_playerID = array.readLittleEndianShort( 51 );
+        m_team = array.readLittleEndianShort( 53 );
+        m_wins = array.readLittleEndianShort( 55 );
+        m_losses = array.readLittleEndianShort( 57 );
+        m_identTurretee = array.readLittleEndianShort( 59 );
+        m_flagsCarried = array.readLittleEndianShort( 61 );
+        m_hasKOTH = array.readByte( 63 );
     }
 
-    public int getTeam(){
+    public short getTeam(){
 
         return m_team;
     }
 
-    public int getWins(){
+    public short getWins(){
 
         return m_wins;
     }
 
-    public int getLosses(){
+    public short getLosses(){
         
         return m_losses;
     }
 
-    public int getHasKOTH(){
+    public byte getHasKOTH(){
 
         return m_hasKOTH;
     }
 
-    public int getShipType(){
+    public byte getShipType(){
 
         return m_shipType;
     }
@@ -78,22 +78,22 @@ public class PlayerEntered extends SubspaceEvent {
         return m_killPoints;
     }
 
-    public int getPlayerID(){
+    public short getPlayerID(){
 
         return m_playerID;
     }
 
-    public int getAcceptsAudio(){
+    public byte getAcceptsAudio(){
 
         return m_acceptsAudio;
     }
 
-    public int getFlagsCarried(){
+    public short getFlagsCarried(){
 
         return m_flagsCarried;
     }
 
-    public int getIdentTurretee(){
+    public short getIdentTurretee(){
 
         return m_identTurretee;
     }
