+++ twcore/twcore/core/TurretEvent.java
@@ -3,22 +3,22 @@ package twcore.core;
 
 public class TurretEvent extends SubspaceEvent {
 
-    private int         m_attacher;
-    private int         m_attachee;
+    private short         m_attacher;
+    private short         m_attachee;
     private boolean     isAttach; //True if attaching, false if detaching
 
     public TurretEvent( ByteArray array ){
-        m_attacher = (int)array.readLittleEndianShort( 1 );
-        m_attachee = (int)array.readLittleEndianShort( 3 );
+        m_attacher = array.readLittleEndianShort( 1 );
+        m_attachee = array.readLittleEndianShort( 3 );
 
         isAttach = ( m_attachee == -1 ? false : true );
     }
 
-    public int getAttacherID() {
+    public short getAttacherID() {
         return m_attacher;
     }
 
-    public int getAttacheeID() {
+    public short getAttacheeID() {
         return m_attachee;
     }
